fix: restructure code
