+++ .env
@@ -1,1 +1,2 @@
 DB_URL="host=localhost user=postgres password=mysecretpassword dbname=postgres port=5432 sslmode=disable"
+PORT=":8090"
