+++ cmd/main.go
@@ -3,8 +3,9 @@ package main
 import (
 	"fmt"
 	"log"
+	"os"
 
-	secretnote "github.com/codescalersinternships/secret-note-api-spa-nabil/api/secretnotehandler"
+	secretnote "github.com/codescalersinternships/secret-note-api-spa-nabil/api"
 	db "github.com/codescalersinternships/secret-note-api-spa-nabil/internal/db/migrate"
 	"github.com/joho/godotenv"
 )
@@ -14,10 +15,10 @@ func main() {
 	if err != nil {
 		log.Fatal("error loading .env file")
 	}
-	DB := db.Init()
+	DB := db.Init(os.Getenv("DB_URL"))
 	r := secretnote.NewServer(DB)
 	// Listen and Server in 0.0.0.0:8090
-	err = r.Start(":8090")
+	err = r.Start(os.Getenv("PORT"))
 	if err != nil {
 		log.Fatal("server run error:", err)
 	}
