+++ docker-compose.yml
@@ -14,10 +14,10 @@ services:
     container_name: secretnote
     build:
       context: ./
-      dockerfile: DockerFile
+      dockerfile: Dockerfile
       target: secretnoteserver
     ports:
-      - "8091:8090"
+      - "8090:8090"
     env_file:
       - .env 
     depends_on:
