+++ internal/db/models/models.go
@@ -4,6 +4,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
+	"gorm.io/gorm"
 )
 
 type User struct {
@@ -23,4 +24,31 @@ type Note struct {
 	User          User      `gorm:"foreignKey:UserID;constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
 	ExpireAt      time.Time `json:"expiredat"`
 	CreatedAt     time.Time `json:"createdat"` // Automatically managed by GORM for creation time
+}
+
+func (user *User) CreateUser(db *gorm.DB) {
+	db.Create(user)
+}
+
+func (user *User) FindByEmail(email string, db *gorm.DB) *gorm.DB {
+	return db.Where("email = ?", email).First(&user)
+}
+func (user *User) FindAllUserNotes(db *gorm.DB) []Note {
+	notes := []Note{}
+	db.Where("user_id = ?", user.ID).Find(&notes)
+	return notes
+}
+func (note *Note) CreateNote(db *gorm.DB) {
+	db.Create(note)
+}
+func (note *Note) FindByID(id uuid.UUID, db *gorm.DB) *gorm.DB {
+	return db.Where("id = ?", id).First(&note)
+}
+
+func (note *Note) Update(db *gorm.DB) {
+	db.Save(&note)
+}
+
+func (note *Note) Delete(db *gorm.DB) {
+	db.Delete(&note)
 }
